requirements.txt
Django==4.2.5
djangorestframework==3.14.0
requests==2.31.0
urllib3==2.0.4
